hw_many/messenger/requirements.txt
amqp==2.6.1
asgiref==3.3.4
billiard==3.6.4.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-elasticsearch-dsl==7.2.0
django-redis==5.0.0
django-sslserver==0.22
django-timezone-field==4.1.2
django-utils-six==2.0
djangorestframework==3.12.4
elasticsearch==7.13.1
elasticsearch-dsl==7.3.0
flower==0.9.7
gunicorn==20.1.0
humanize==3.7.0
idna==2.10
kombu==4.6.11
oauthlib==3.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
